+++ app/components/canvas/elements/preview/placeholderBalls.module.css
